Updated persistent tags to become inactive again. Note that persistent tags use local variables rather than context.
