+++ src/components/TagCloud/TagCloud.js
@@ -2,9 +2,7 @@ import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import Chip from '@material-ui/core/Chip'
 import AddCircleIcon from '@material-ui/icons/AddCircle'
-import DeleteIcon from '@material-ui/icons/Delete'
-import ChipInput from 'material-ui-chip-input'
-import { UserStateContext } from '../../context/UserContext'
+import HighlightOffIcon from '@material-ui/icons/HighlightOff'
 
 const useStyles = makeStyles(theme => ({
 	root: {
@@ -19,7 +17,7 @@ const useStyles = makeStyles(theme => ({
 
 function setDeleteIcon(chip) {
 	if (chip.active === true) {
-		return <DeleteIcon />
+		return <HighlightOffIcon />
 	}
 
 	return <AddCircleIcon />
@@ -27,35 +25,35 @@ function setDeleteIcon(chip) {
 
 export default function ChipsArray() {
 	const classes = useStyles()
-	// [
-	// 	{ key: 0, active: true, color: 'primary', label: '#gasleak' },
-	// 	{ key: 1, active: true, color: 'primary', label: '@coned' },
-	// 	{ key: 2, active: true, color: 'primary', label: '#help' },
-	// ]
+	const [chipData, setChipData] = React.useState([
+		{ key: 0, active: true, color: 'primary', label: '#gasleak' },
+		{ key: 1, active: true, color: 'primary', label: '@coned' },
+		{ key: 2, active: true, color: 'primary', label: '#help' },
+	])
 
 	// this arrow function returns ANOTHER arrow function
 	// so the handler has to be passed an argument
 	// handleDelete(chip) gets you a function to delete the chip
 	// so it isn't executed immediately (as might be the case if
 	// the function returned a primitive like a number or object or something
-	const handleDelete = (tagToDelete, context) => () => {
-		context.onHashtagsChange(hashtags =>
-			hashtags.map(hashtag => {
-				if (hashtag.key === tagToDelete.key) {
-					if (hashtag.active === true) {
+	const handleDelete = chipToDelete => () => {
+		setChipData(chips =>
+			chips.map(chip => {
+				if (chip.key === chipToDelete.key) {
+					if (chip.active === true) {
 						return {
-							...hashtag,
+							...chip,
 							active: false,
 							color: '#b3b3b3',
 						}
 					}
 					return {
-						...hashtag,
+						...chip,
 						active: true,
 						color: 'primary',
 					}
 				}
-				return hashtag
+				return chip
 			}),
 		)
 	}
@@ -64,31 +62,18 @@ export default function ChipsArray() {
 	// to reactivate it.
 	return (
 		<div className={classes.root}>
-			<UserStateContext.Consumer>
-				{context => {
-					context.settings.hashtags.map(chip => {
-						return (
-							<Chip
-								key={context.settings.hashtags.key}
-								label={context.settings.hashtags.label}
-								onDelete={handleDelete(chip, context)}
-								className={classes.chip}
-								color="primary"
-								deleteIcon={setDeleteIcon(chip)}
-							/>
-						)
-					})
-
-					return (
-						<ChipInput
-							defaultValue={context.state.settings.hashtags}
-							label="Persistent Settings Hashtags"
-							onChange={handleDelete}
-							disableUnderline
-						/>
-					)
-				}}
-			</UserStateContext.Consumer>
+			{chipData.map(chip => {
+				return (
+					<Chip
+						key={chip.key}
+						label={chip.label}
+						onDelete={handleDelete(chip)}
+						className={classes.chip}
+						color={chip.color}
+						deleteIcon={setDeleteIcon(chip)}
+					/>
+				)
+			})}
 		</div>
 	)
 }
